Remove hooks Refactor GestureMonitor
