+++ src/index.ts
@@ -51,8 +51,6 @@ export type {
   NativeViewGestureHandlerProps,
 } from './handlers/NativeViewGestureHandler';
 export { GestureMonitor } from './handlers/gestures/GestureMonitor';
-export { useGesture } from './handlers/gestures/useGesture';
-export { useAnimatedGesture } from './handlers/gestures/useAnimatedGesture';
 export { GestureObjects as Gesture } from './handlers/gestures/gestureObjects';
 export { NativeViewGestureHandler } from './handlers/NativeViewGestureHandler';
 export type {
